ECWID-142216 - add generic method to use searchOrders
