+++ src/main/kotlin/com/ecwid/apiclient/v3/impl/OrdersApiClientImpl.kt
@@ -2,9 +2,13 @@ package com.ecwid.apiclient.v3.impl
 
 import com.ecwid.apiclient.v3.ApiClientHelper
 import com.ecwid.apiclient.v3.OrdersApiClient
+import com.ecwid.apiclient.v3.dto.category.request.CategoriesSearchRequest
+import com.ecwid.apiclient.v3.dto.category.result.CategoriesSearchResult
+import com.ecwid.apiclient.v3.dto.common.PartialResult
 import com.ecwid.apiclient.v3.dto.order.request.*
 import com.ecwid.apiclient.v3.dto.order.result.*
 import com.ecwid.apiclient.v3.responsefields.AS_SEQUENCE_SEARCH_RESULT_REQUIRED_FIELDS
+import kotlin.reflect.KClass
 
 internal class OrdersApiClientImpl(
 	private val apiClientHelper: ApiClientHelper
@@ -12,6 +16,16 @@ internal class OrdersApiClientImpl(
 
 	override fun searchOrders(request: OrdersSearchRequest) =
 		apiClientHelper.makeObjectResultRequest<OrdersSearchResult>(request)
+
+	override fun <Result : PartialResult<OrdersSearchResult>> searchOrders(
+		request: OrdersSearchRequest,
+		resultClass: KClass<Result>
+	): Result {
+		return apiClientHelper.makeObjectPartialResultRequest(
+			request = request,
+			resultClass = resultClass,
+		)
+	}
 
 	override fun searchOrdersAsSequence(request: OrdersSearchRequest) = sequence {
 		var offsetRequest = request.copy(
